two small patches
git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@1122 a592a061-630c-0410-9148-cb99ea01b6c8

+++ src/combox.C
@@ -142,7 +142,7 @@ bool Combox::select_text(string const & t)
 }
 
 
-void Combox::select(int i)
+void Combox::select(size_t i)
 {
 	if (!browser || !button) return;
 	if (i > 0 && i <= fl_get_browser_maxline(browser)) sel = i;
